submissions/my-github-id/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "banking-client"
version = "0.1.0"
description = "Modern Python banking client for Julius Baer challenge"
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "--import-mode=importlib"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
target-version = "py312"
line-length = 100
